+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Bugfixes
 
 - [Multiple] Fix ckan version command and update copyrights (#4347 by: HebaruSan)
+- [GUI] Suppress exceptions on setting SplitterDistance (#4350 by: HebaruSan)
 
 ### Internal